[ISSUE #20]Abstracting the server and client and implementation
